+++ scripts/addons/cam/utilities/addon_utils.py
