+++ scripts/addons/cam/utilities/operation_utils.py
