+++ scripts/addons/cam/utilities/orient_utils.py
