+++ scripts/addons/cam/utilities/shapely_utils.py
