--- scripts/addons/docs/cam.opencamlib.rst
@@ -1,29 +1,0 @@
-cam.opencamlib package
-======================
-
-Submodules
-----------
-
-cam.opencamlib.oclSample module
--------------------------------
-
-.. automodule:: cam.opencamlib.oclSample
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-cam.opencamlib.opencamlib module
---------------------------------
-
-.. automodule:: cam.opencamlib.opencamlib
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: cam.opencamlib
-   :members:
-   :undoc-members:
-   :show-inheritance:
